requirements.txt
pinecone>=4.0.0
python-dotenv>=1.0.0
pinecone-plugins>=0.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1